increasing maximum allowed serverside char count
This change is in response to Zendesk ticket
https://govuk.zendesk.com/agent/#/tickets/43023.

Before this change, there is a 1200 char limit on certain form fields
within both the frontend (JS, templates) and the backend (models).
However, because some browers treat "\r\n" as 1 char, it is possible
to create a submission that the frontend allows but fails the server-
side validation.

To work around this, this change bumps the model validations to 1250
chars.