FOI requests raised by 'support' app
This change moves the responsibility of raising FOI requests from
this app to the 'support' app. Instead of connecting to Zendesk,
this app posts to a 'support' endpoint (which is protected by
signon).

This is the first step in having a more frontend-backend relationship
between the 'feedback' and 'support' apps, and will allow better
response times for FOI requests because the round-trip to 'support'
is much shorter than to Zendesk (and 'support' raises the Zendesk
ticket via a background queue).