remove unnecessary sectional context
Before this change, the named contact form allowed the user to select
sections of the site as an answer to "what's it about". While this info was
being captured, it wasn't being used for anything except routing tickets about the
Service Manual to the appropriate Zendesk group.

The change removes the section option; routing of Service Manual tickets
will be done on the basis of the referrering page matching /service-manual/.