remove FOI request link from index page
since the renaming of 'feedback' to 'contact' in the footer and the
introduction of the contact index page, many users are submitting FOI
requests that aren't legitimate FOI requests, and should really be
coming through the contact form instead.

The FOI request form is still accessible through the /help pages.