more accurate description of the help pages
currently, the description of the help pages promises more than the /help
pages deliver. Analytics show that almost as many users click through to the
/help pages as go to the contact form.

This updates the description on the link to reflect
the actual content on that page. The copy itself closely follows the explanation
on the /help page.