Save URL when coming direct to the contact form
This change will automatically set the "A specific page" radio button
on the public contact form if the referrer isn't the interstitial (/contact)
page.

This is useful whenever there is a direct link to the public form (eg from
`www.gov.uk/service-manual`); in those cases, this change will mean that the
specific page URL will be captured without user intervention. For users coming
from `/contact`, they will continue having 'What's it to do with?' default to
'The whole site'.