Save the user's journey to the contact form
The 'Specific page' field on the /contact/govuk form auto-populates
with the previous page (ie the one the user followed the link from to
get there). Originally, this would usually have been the page the user
had been on when they spotted the issue they want to raise however,
since we introduced the interstitial /contact page almost all traffic
comes via this page
- this means almost all named feedback that comes through to 2nd line
relating to content issues lists the path as /contact, not the page the
user is feeding back on.

* Capture the original page the user started on, using the referrer to
the /contact page, and saving it to a cookie
* Pre-populate the specific page value with the value from the cookie
if it’s been set