Bump govuk-lint, re-configure, fix outstanding violations
A newer lint always brings new violations so we fix them, hence
the noise in the diff.  We add the --rails option to get rails
specific cops to the linting, and we also add local config asking
the linter to ignore automatically generated files spec/spec_helper
and spec/rails_helper as we want to keep the diff on those from the
default versions as small as possible.