Add model and controller for assisted digital help with fees
This is a no-op in this commit, we're just adding the boilerplate for
routing the requests from the new form in frontend and validating it.

Like service feedback, a failed validation won't re-render the form
to the user because the original form lives in frontend so it'd be
too hard.

A future commit will add the code to persist the values sent to a
backing store (mostly likely google docs).