Add google spreadsheet backing store for assisted digital
We use the google-api-client gem to power the store.  When "saving" the
assisted digital help with fees feedback we append the data to a
spreadsheet.  The spreadsheet is provided in a ENV var (so will come
from puppet eventually), and the auth details are currently using the
"application default" strategy.  This means we generate a service account
in the google developer console and store the keys in a JSON file on disk.
The email address of the service account must be given explicit access to
the spreadsheet (as it does not reside on the digital.cabinet-office.gov.uk
domain so it won't get "default domain" access).

This should mean we can have different spreadsheets per deployed environment
and different service accounts too.