Authenticate with google via env vars not JSON files
The only values we need from the JSON file are client_email and
private_key and the googleauth gem supports setting these via
environment variables (GOOGLE_CLIENT_EMAIL and GOOGLE_PRIVATE_KEY
respectively).  This means we don't need to store the JSON key file on
the deployed servers and all configuration can be done via puppet.

There is a bug in the version of the googleauth gem we use (0.5.1) that
means the `Google::Auth.get_default_credentials` method doesn't work when
using it via env vars only, so we have to reach into the API and use some
of the underlying implementation instead.  It's been fixed in master (see:
https://github.com/google/google-auth-library-ruby/commit/62a8d41bc4e3274f9f3b7ceab63a040912ee2bdc)
but not released.  When a version is released we can change back to using
the `get_default_credentials` version, although we do also have to set
another environment variable (GOOGLE_ACCOUNT_TYPE set to "service_account")
to tell it what sort of credentials we want it to make, which we can avoid
by reaching into the API.