Attempt at more effective spam honeypot styling
Attempt to reduce the number of spam contact requests that are occurring
in our zendesk account.

This commit moves the style of these honeypot fields to an external
file. It avoids the more simple inline styling that existed as this may
be more easily detected by spambots
(http://ux.stackexchange.com/questions/52916/what-is-the-best-way-to-hide-a-honeypot-captcha)

Also:
  - Added a label and aria role to the honeypot field so that screenreaders will
  indicate that it's not to be filled in.