Rework AssistedDigitalHelpWithFeesFeedback
This now takes more options and conditionally validates them based on
their values.  For example, if "assistance_received" is "no" then we don't
care about the values sent for "assistance_provided_by" or
"assistance_received_comments".  As before, we can't re-present the form to
the user if it's filled in incorrectly, so we just say "thanks" anyway.