Introduce a MultiTicket model
This acts as a wrapper to multiple ticket types and performs all actions
on all the tickets before returning a combined value.  E.g. if any of the
tickets are spam the whole multiticket is spam, but all of them have to be
valid for the whole multiticket to be valid.