Make submitting assisted digital feedback act like service feedback
We still want the service feedback to go into the standard performance
platform feedex data and the easiest way to do that is to use the
ServiceFeedback ticket type.  We do want to keep sending
the full assisted digital data to the google spreadseeht and the easiest
way of doing that is to keep using the
AssistedDigitalHelpWithFeesFeedback ticket type.  To use both we use the
new MultiTicket type and do both.