Introduce EmailSurvey and EmailSurveySignup
EmailSurvey is a value object with a simple all/find interface to find
instances by id.  These value objects will contain the details of the
surveys that are currently running.  Like their counterparts from static
in surveys.js we have an id, a url and start/end times.  This does mean
there will be some duplication between static and feedback for deploying
new email surveys - we may want to investigate minimising this.

EmailSurveySignup is a "ticket" model that we can use in a controller to
convert the params from static into an email going to the user via notify.
We don't extend from Ticket because we don't need all the default
behaviour with val, url, referrer, user_agent and so-on.  Note that we've
not tried to plug notify into the app yet, but our intention is that it
will go inline into the EmailSurveySignup#save method just like other
tickets that write to support-api or google spreadsheets.