Saving an EmailSurveySignup triggers the SurveyNotifyService
Maybe it'd be good to put this in a worker rather than the
request/response cycle, but none of the other "ticket" types do this so
I think we'll be ok.

Noe that asking GOV.UK Notify to send the email is asynchronous too.  The
success response tells us that notify has queued it for delivery *at some
point* and we are supposed to check using the API about the eventual state.
I figure that this app doesn't need to do that, it can be done manually via
the GOV.UK Notify dashboard by intersted parties.