Add the controller and route for email survey signup
The controller does extend from the existing ContactController (which is
why the EmailSurveySignup doesn't extend from Ticket, but does present the
same API).  Because the form in static is submitted via AJAX we provide
xhr responses that return JSON, but leave the standard behaviour for HTML
requests.  Like other feedback tickets if the request is made via HTML and
is invalid there's not much we can do because the form lives in static, not
feedback so we just let folk think it was a success.

We also extract the exception logging from the application_controller's
unable_to_create_ticket_error method so we can use it in our xhr error
response.  Otherwise we wouldn't see the notify errors in errbit, or our
statsd logging and we probably want to.