Allow .js routes for email signup and treat as ajax
It seems that our deployed infrastructure strips off the
`X-Requested-With` header that rails' `request.xhr?` method uses
to work out if a given request is ajax or not.  To counteract this
we allow a `.js` version of the route for email signup and treat
that as an ajax request as well as those with the `X-Requested-With`
header and value that `.xhr?` uses.