Only provide survey_url to notify for template personalisation
Turns out that notify will error if the personalisation details contain
extra keys or missing keys.  Now that the template only wants survey_url
we need to stop sending survey_name.

Added some notes about this to the readme as it means we need to take
extra care about the params and templates.