Set correct identifier and dates for first email survey
The id needs to be `govuk_email_survey_t01` because the previous version,
`education_email_survey` implied it was about education when in fact it
was explicitly to exclude education.  We also change the dates to run from
the 8th to the 13th.  The frontend counterpart is actually 2 surveys; one
that runs from the 8th to 1pm on the 10th, and another that runs from 1pm
on the 10th to the end of the day on the 13th.  The only difference is the
frequency so our backend can treat them as the same survey.