Use separate survey definitions for the Thursday and Weekend surveys
The user research team became concerned that they wouldn't be able to tell
which users were signing up to the survey via the 1-in-1 version running
on Thursday->Friday lunch and which were signing up via the 1-in-2 version
running from Friday lunch to Monday eve.  To that end we want different
identifiers (which will be used in events in the frontend) and different
survey urls.