Add GA client ID to email surveys
For: https://trello.com/c/xDdnial5/220-clientid-pushed-into-email-surveys-as-custom-data

In order to track the user journey when a user completes an EMAIL survey, we can use the client ID to identify a session for a user and record all the pages he had visited before arriving to our survey. This will provide more detailed data about our user's browsing habits and will help to separate the data out, since the info collected on surveys is anonymous.