Include referrer from posted params if present
The forms in frontend that are used to submit service feedback have the
referrer stored in a hidden field added by JS with the name `referrer`.
It is not present in the `service_feedback` params which all our existing
tests assumed.  The outcome of this is that none of the submitted service
feedback has a useful referrer value stored; it's either nil, or since
we merged #206, filled in with the request referer value which will be the
page that the service feedback form was presented on.  That's not a useful
value for the referrer in these cases because we know what page the form
is on, it's the /done page - we want to know where they came from before
they got to the /done page.

We change the contact controller to fetch the referrer from either the
service_feedback hash, or the posted params, or the request.referer (in
that order of preference).