Use hybrid for cookie serialization
Earlier all the cookies were serialized using Marshal library of Ruby.
The marshalling of cookies can be unsafe because of the possibility of
remote code execution vulnerability.

After this, all the existing marshalled cookies will be migrated to
:json format properly and in the future upgrade of Rails, we can safely
change the config from :hybrid to :json which is the default and safe
value of this config.

See: http://blog.bigbinary.com/2014/12/23/migrating-existing-session-cookies-while-upgrading-to-rails-4-1-and-above.html