Update cookie serialiser to json format
Earlier all the cookies were serialized using Marshal library of Ruby.
The marshalling of cookies can be unsafe because of the possibility of
remote code execution vulnerability.

In a previous commit we migrated cookies using the hybrid serialiser
which now needs to use the json serialiser to finalise the change