Use feedback component
This replaces the "Is there anything wrong?" link with the new feedback
component. This component asks the user if the page is useful and
reports the results to Google Analytics.

See https://github.com/alphagov/govuk_publishing_components/pull/163
for the component.

This removes any testing of the feedback component specifics. This just
tests the form, which is already done in the components gem.

https://trello.com/c/eKtXGtBD