Add a new email survey for the yes/no feedback banner
When you click "No" in the feedback component you're asked to fill in
your email address so we can send you a survey. The survey link we
currently send (https://www.smartsurvey.co.uk/s/gov-uk) is the default
survey that is also present in the blue banner. This makes analysing
the data very difficult.

This adds a new EmailSurvey type with a new URL that Katie provided.
After this is merged & deployed we can update the feedback component to
use `footer_satisfaction_survey` instead of `user_satisfaction_survey`
for the `survey_id`[1].

[1]:
https://github.com/alphagov/govuk_publishing_components/blob/94400e31e6a
45e0b1f55d3d43afee5e0575db2d9/app/views/govuk_publishing_components/comp
onents/_feedback.html.erb#L120