Notify Sentry about invalid survey submission
When a user answers "No" to the "Is this page useful?" at the bottom of
the page, they can fill in an email address. We're seeing that the
endpoint that receives the email address
(`/contact/govuk/email-survey-signup`) is returning 422 quite a lot -
about 800 in the last week.

This commit adds a notification to Sentry so we can figure out why
these requests are being rejected. Currently this info isn't logged, so
we haven't got a good idea.