Remove “robot” from the honeypot contact form code
We tried changing the honeypot field name from “val” to “giraffe” in a previous iteration of testing spam improvements. That didn’t have any impact.
I was suggested that robots might be ignoring it as the label in the HTML for that field is “ignore if you’re not a robot”, and the robots were catching “robot”.

Single: @matthewculluum-gds