Read remaining characters when threshold reached
Update to PR #267

Ideally we'd only want to enable screenreader character count announcements only when a certain threshold is reached.

This PR amends the character counter javascript so that it removes aria- attributes when users start typing (leaving the initial limit announcement intact)

and attaches attributes only when 90% character count threshold is reached.