Make tested methods public
We currently expect that
- `improvement_comments`, `assistance_received_comments` and
  `assistance_improvement_comments` in `assisted_digital_feedback.rb`
- `improvement_comments` in `service_feedback.rb`
- `link` in `contact_ticket.rb`
all return `nil` instead of empty strings when not present.

This is achieved by doing (considering only one of them as an example)
```
attr_writer :improvement_comments

...

def improvement_comments
  @improvement_comments.present? ? @improvement_comments : nil
end
```
as oppose to
```
attr_accessor :improvement_comments
```
that would return an empty string when not present.

I don't have much context on why want `nil` as opposed to empty string,
but we have tests in place for this
```
it { is_expected.to allow_value(nil).for(:improvement_comments) }
```
so I assume it is an expected and wanted behaviour.

When bumping `shoulda-matchers` from 3.0.1 to 4.1.2, some of these tests
started failing. This is because some of the methods described are
private. Making them public as `link` solves the failing tests without
having to change the behaviour of the app.