Disable CSRF protection
As the feedback form is cached in the CDN we can't use cookie-based
sessions for CSRF protection so disable it by default but allow it
in the test environment to ensure it doesn't get re-enabled by mistake.

CSRF protection was accidentally re-enabled during the upgrade from
Rails 5.1 to 6.0 because the setting was added to ActionController::Base
and the protect_from_forgery line removed from ApplicationController.