Rework readme for consistency and GOV.UK docker references
My initial motivation for this update was to make the technical
documentation consistent in the references to GOV.UK docker and remove
the outdated information for bowler and startup.sh.

In the process I noticed the technical documentation of feedback types
supported wasn't particularly sensical and seemed to be repeated later
in this document and in additional documentation. Therefore, I removed
the references to this and made the links to the other documentation
more prominent.

As part of this I made the rest of the doc a bit more conventional with
GOV.UK README consistency with some minor ordering changes and the
removal of unusual backticks.