Add AccessibleFormatRequest controller and views
This adds a controller and views for an accessible format request. The controller reads from a YAML file to access a list of questions.

Page numbers are passed via parameters to determine which question to render.

The view contains a hidden submission element to allow previous answers to persist in subsequent submissions.

Permitted question parameters are read from the YAML file
The initial request will be made with additional information including parent content_id and the requested attachment_id.