Change to legend
GOV.UK Contact page was flagged as having a confusing heading structure. 

"Contact GOV.UK" is h1 
"What's it to do with?" is a h2 however
"What are the details?, Do you want a reply?" are legends

The ideal solution would be to restructure this page to ask "one question per page" as per the Design System's Guidance[1]

Also it was discovered that nested headings inside legends has been added to the HTML spec, in part, by work from the Design System team[3][4]

However from an accessibility POV - from this update - there seem to be some additional considerations to take into account with JAWS not recognising headings while nested.[5]

One balance, the accessibility team decided to remove the h2 on the page.  This is then more inline with Design System guidance and resolves the confusing structure.

[1]https://design-system.service.gov.uk/patterns/question-pages/#start-by-asking-one-question-per-page

[2]https://design-system.service.gov.uk/get-started/labels-legends-headings/

[3]https://github.com/w3c/html/issues/724
[4]https://github.com/whatwg/html/pull/5090
[5]https://github.com/alphagov/govuk-frontend/issues/1605