Present format details to_notify_params
If the custom_details attribute is present this should be presented
to Notify. Notify require that all "personalisation" keys are present
so a placeholder string is sent if no custom details are provided..

If a submitted format type has an underscore, e.g british-sign-language,
this should be presented as a human-readable string e.g British sign language.