Add error handling for a failed AccessibleFormatRequest
Should the AccessibleFormatRequest be invalid (e.g not all parameters
were provided) or fail to save (e.g the Notify service or PublishingApi
throws an error) we show an error page. The PublishingApi may fail to return
a content item if an invalid content_id is provided, or may be unable to
find an attachment if an invalid attachment_id is provided.

For errors we also log the error using GovukError, adapting a
`unable_to_create_ticket_error` method provided by the ApplicationController to catch
Notify errors. This is already configured for PublishingApi errors.

If possible (i.e the information has been provided by PublishingApi) we will
render a link on the error page to return the user to the original document page.

Co-authored-by: Jon Kirwan <jon.kirwan@digital.cabinet-office.gov.uk>