Add error handling for a failed AccessibleFormatRequest
Should the AccessibleFormatRequest be invalid (e.g not all parameters
were provided) or fail to save (e.g the Notify service or PublishingApi
throws an error) we show an error page.

For errors we also log the error using GovukError, adapting a
`unable_to_create_ticket_error` method provided by the ApplicationController to catch
Notify errors. This is already configured for PublishingApi errors.

If possible (i.e the information has been provided by PublishingApi) we will
render a link on the error page to return the user to the original document page.