"Request accessible document" form - add layout and components
This adds a route and controller with placeholder views for an accessible format request. The controller reads from a YAML file to access a list of questions.

Page numbers are passed via parameters to determine which question to render.

The view contains a hidden submission element to allow previous answers to persist in subsequent submissions.

Permitted parameters are read from the YAML file

When all questions have been answered, the final view will instead make a POST request to submit all params the unimplemented #create method.