Display error messages if required input is missing
If a user fails to click a button or enter text in a required field,
this commit ensures that the same page will be reloaded and a helpful
error message displayed.

In a before action the controller will check the inputs of the
previous question asked and, if any keys with an associated error
message are not present in the params, will set these error messages
in the flash and repost a form using the previous question number.

This commit makes use of the "repost" gem
https://rubygems.org/gems/repost/versions/0.3.8