Comment out specify puma pid location
Setting this pidfile requires a tmp/pids directory to be created
otherwise puma will error. This is a pain point for us as in CI puma is
used to run the JavaScript tests.

An alternate fix for this would be to use a git .keep in tmp/pids but
the convention in existing GOV.UK apps is not to set this config.