Add before_action to ensure document and attachment are valid
Previously we were implicitly testing that the content_item/document and the
attachment were both valid by relying on them when rendering views.

The new start page does not exercise the attachment. This before_action
ensures that both necessary elements are valid and available at every stage
of the journey other than "request_sent" where we no longer mind if
the content item cannot be retrieved.