Route completed transactions to /done/completed-transaction-name
The completed_transaction done pages are being moved from Frontend to Feedback so that model validations can be rendered, as they are defined within Feedback.

The completed_transaction done pages are either standard surveys (service_feedback) such as www.gov.uk/done/vehicle-tax or assisted digital satisfaction surveys (assisted_digital_feedback) such as https://www.gov.uk/done/register-flood-risk-exemption. There are only three assisted digital satisfaction surveys.

This routing approach has been lifted from the Frontend application by which there is a routing constraint on /done/completed-transaction paths. This looks up the content item, checks that it is a completed_transaction and caches it. If it is a completed_transaction then it is routed to a controller action. This routing/constraint will be removed from the Frontend application.

In the Frontend application, the same routing constraint is used for all of the done pages
and then different partials are rendered depending on whether the form is an Assisted Digital survey. This would be problematic here due to both forms needing to post to different controllers to render the correct model validations. If the same routing constraint was used then the application would struggle to determine which controller action would re-render the form. Therefore, the approach is different and the request is routed to the correct controller from the outset in Routes.rb.

Given there are only three assisted digital routes they have been manually defined. This was also due to it being difficult to think of a regex that would be able to identify these pages from the other completed_transaction done pages.