Get completed_transaction content item data
The content on the forms (such as the title and whether to display a promo box) is rendered dynamically from the content item. Therefore, a call needed to be made to Content Store to retrieve the correct content to render.

When the form was in Frontend, it submitted the data to the Feedback app, so the controllers already existed. They all inherit from ContactController, so the #new action was overwritten in the subclasses to render @Publication. @Publication stores the presented content item (fetched from the Content Store if it isn’t already cached). Fetching and formatting the content item is handled by ServiceFeedbackHelper and ContentItemPresenter.