Exclude from throttling
The application kept throwing 429 errors. This was due to rate limiting set in RackAttack. The post request needs to make another request to content store when re-rendering the page (if it isn't cached) in order to re-render the content specific to the service on the view. It has therefore been excluded.