Service Feedback Create action for submitting form
If there are validation errors then the form re-renders. Fortunately the validations were already in place. When the form was in the Frontend app, data was submitted to this Feedback application, but the code that did the validation was purposefully not run within Feedback as it was not possible to render the validations back onto the Frontend app. Moving the form into feedback will mean that the pre-existing model validations can be rendered.

There is no database in this application, so the form values are stored as instance variables. Instance variables are also generated to store the error messages for each field and the error-id so that the message summary links to the correct view component. This approach was taken as opposed to using Flash messages as it enabled making use of the existing validation errors.