New and Create actions for assisted_digital_feedback
If there are validation errors then the form re-renders. Fortunately the validations were already in place. When the form was in the Frontend app, data was submitted to this Feedback application, but the code that did the validation was purposefully not run within Feedback as it was not possible to render the validations back onto the Frontend app. Moving the form into feedback means that the pre-existing model validations can now be rendered.

There is no database in this application, so the form values are stored as instance variables. Instance variables are also generated to store the error messages for each field and the error-id so that the message summary links to the correct view component. This approach was taken as opposed to using Flash messages as it enabled making use of the existing validation errors.

There is some duplication in the error message ivar generating methods between this controller and service_feedback_controller. There was an attempt to move them into a helper but the view didn’t recognise them. They didn’t feel generic enough to move to the parent class so they have been duplicated.

The assisted_digital_feedback form displays extra parts of the form depending on what the user selects. The assistance_satisfaction_rating and assistance_improvement_comments were being used by multiple parts of the form which caused the validation errors to render in the wrong place. Therefore fields called assistance_satisfaction_rating_other and assistance_improvement_comments_other were added to prevent this issue. The order of the validation errors was also changed at points so that the more user friendly error displayed first.