Append javascript_enabled form field
Append hidden javascript_enabled form field to service_feedback and assited_digital_feedback forms.
The referer wasn't passed through, so far it has just been the /done page URL which is set
in the controller. This seems ok as we would expect the page to be visited as a result of completing a
transaction.