Add Photo ID Promo
Reimplements a change that was added into the Frontend application recently [1]. The change therefore needed to be re-implemented in here. It adds a promo box for bringing photo ID to vote. This change is temporary and these promo boxes will need to be removed on 04/05/2023 according to [2]. The author of the change introduced a nice refactor for rendering the promo box and translations.

[1](https://github.com/alphagov/frontend/pull/3531/files)
[2](https://github.com/alphagov/frontend/pull/3544)