Fix naming issue with slug and base_path
It was realised that the naming in the completed_transaction code caused some confusion as there were places where what is actually the 'base_path' is referred to as the 'slug'. In reality we need the 'base_path' for rendering the form from the content item and we need the 'slug' to be passed through the form as a hidden field and onto the Support API..

Using 'slug' in the routing constraint etc was due to following the same pattern as was used in the Frontend application, which is where the '/done/transaction' pages were previously rendered from.

[The PR to switch the rendering of completed_transactions is here](https://github.com/alphagov/feedback/pull/1601/).