Make Zendesk API request to 'users' endpoint
This will be used to query the 'suspended' status of users. But
before we do that, we have to make a bunch of edits to the
existing tests to ensure they continue to work, so let's just
make the call and do nothing with the value at this stage. See
next commit to see the value being used.